+++ src/pl/testuj/operations/CModuloOperation.java
@@ -4,10 +4,14 @@ package pl.testuj.operations;
  * Created by krzysztoffonal on 10/09/16.
  */
 public class CModuloOperation extends COperation {
-  @Override public double process(double firstArg, double secondArg) {
-    double r = firstArg;
+  /*
+   * Assumption: Choose scenario for minus args - if you want to calculate it properly
+   * or always return 0
+   */
+  @Override public int process(int firstArg, int secondArg) {
+    int r = firstArg;
 
-    while (r>=secondArg)
+    while (r >= secondArg)
     {
       r = r - secondArg;
     }
