+++ src/pl/testuj/operations/CPowerOperation.java
@@ -4,7 +4,7 @@ package pl.testuj.operations;
  * Created by krzysztoffonal on 10/09/16.
  */
 public class CPowerOperation extends COperation {
-  @Override public double process(double firstArg, double secondArg) {
+  @Override public int process(int firstArg, int secondArg) {
     return firstArg * firstArg;
   }
 }
